+++ CHANGELOG.md
@@ -223,6 +223,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 ### Added
 
 - Implement the following traits
+  - From\<T: PrimInt\> for Der (`PrimInt` is `i8` or `u8` or ... `u128` or `isize` or `usize`.)
   - From\<&[u8; N]\> for Contents
   - From\<&[u8; N]\> for &ContentsRef
   - From\<&mut [u8; N]\> for &mut ContentsRef